Refactor test for easier reuse.
